% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_core.R
\name{aggregate_light_unit_infections}
\alias{aggregate_light_unit_infections}
\title{Aggregate light protocol infections by ICU}
\usage{
aggregate_light_unit_infections(unit_light, denominator_light, infection_light)
}
\arguments{
\item{unit_light}{Light protocol unit table.}

\item{denominator_light}{Light protocol denominator table.}

\item{infection_light}{Light protocol infection table.}
}
\value{
Light protocol unit infection aggregate.
}
\description{
Aggregate light protocol infections by ICU
}
