% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_core.R
\name{aggregate_standard_unit_infections}
\alias{aggregate_standard_unit_infections}
\title{Aggregate standard protocol infections by ICU}
\usage{
aggregate_standard_unit_infections(patient_infections, unit)
}
\arguments{
\item{patient_infections}{Patient-infection table.}

\item{unit}{Standard protocol unit table.}
}
\value{
Standard protocol unit infection aggregate.
}
\description{
Aggregate standard protocol infections by ICU
}
