% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_core.R
\name{build_haiicuall2}
\alias{build_haiicuall2}
\title{Build the combined incidence input table}
\usage{
build_haiicuall2(standard_unit_infections, light_unit_infections)
}
\arguments{
\item{standard_unit_infections}{Standard protocol unit infection aggregate.}

\item{light_unit_infections}{Light protocol unit infection aggregate.}
}
\value{
Combined incidence input equivalent to historical \code{haiicuall2}.
}
\description{
Build the combined incidence input table
}
