% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{haiicu_input_files}
\alias{haiicu_input_files}
\title{HAIICU raw input filenames}
\usage{
haiicu_input_files()
}
\value{
Named character vector of expected raw TESSy export filenames.
}
\description{
HAIICU raw input filenames
}
