% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAINETICU-package.R
\docType{package}
\name{HAINETICU-package}
\alias{HAINETICU}
\alias{HAINETICU-package}
\title{HAINETICU: HAI-Net ICU Surveillance Annual Report Workflow}
\description{
Package infrastructure for the HAI-Net ICU annual surveillance workflow.
}
\author{
\strong{Maintainer}: HAI-Net ICU surveillance analysis contributors \email{HAI-Net@ecdc.europa.eu}

}
\keyword{internal}
