% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tables.R
\name{hicu_append_eu_row}
\alias{hicu_append_eu_row}
\title{Append an EU/EEA row to a country/network table}
\usage{
hicu_append_eu_row(
  country_data,
  eu_data,
  country_col = "ReportingCountry",
  eu_label = "EU/EEA"
)
}
\arguments{
\item{country_data}{Country/network data frame.}

\item{eu_data}{One-row EU/EEA data frame.}

\item{country_col}{Name of the country/network column.}

\item{eu_label}{Label to use for the EU/EEA row.}
}
\value{
\code{country_data} with an EU/EEA row appended last.
}
\description{
Append an EU/EEA row to a country/network table
}
