% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_paths.R
\name{hicu_default_data_dir}
\alias{hicu_default_data_dir}
\title{Build default raw data directory path}
\usage{
hicu_default_data_dir(year = hicu_default_year())
}
\arguments{
\item{year}{Character year value.}
}
\value{
Absolute path to data/raw/\if{html}{\out{<year>}}.
}
\description{
Build default raw data directory path
}
\keyword{internal}
