% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_paths.R
\name{hicu_default_output_dir}
\alias{hicu_default_output_dir}
\title{Build default output directory path}
\usage{
hicu_default_output_dir(year = hicu_default_year())
}
\arguments{
\item{year}{Character year value.}
}
\value{
Absolute path to outputs/\if{html}{\out{<year>}}.
}
\description{
Build default output directory path
}
\keyword{internal}
