% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_paths.R
\name{hicu_default_year}
\alias{hicu_default_year}
\title{Get default reporting year}
\usage{
hicu_default_year()
}
\value{
Character scalar year.
}
\description{
Get default reporting year
}
\keyword{internal}
