% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_paths.R
\name{hicu_find_project_root}
\alias{hicu_find_project_root}
\title{Find project root directory}
\usage{
hicu_find_project_root()
}
\value{
Absolute path to the project root.
}
\description{
Find project root directory
}
\keyword{internal}
