% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{hicu_label_country_values}
\alias{hicu_label_country_values}
\title{Label country/network values}
\usage{
hicu_label_country_values(
  data,
  country_col = "ReportingCountry",
  labels = hicu_country_labels()
)
}
\arguments{
\item{data}{Data frame containing a country/network column.}

\item{country_col}{Name of the country/network column.}

\item{labels}{Named character vector of labels.}
}
\value{
\code{data} with labelled country/network values where labels are known.
}
\description{
Label country/network values
}
