% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{hicu_report_var_labels}
\alias{hicu_report_var_labels}
\title{Report table display labels}
\usage{
hicu_report_var_labels()
}
\value{
Named character vector mapping report column names to labels.
}
\description{
Report table display labels
}
