% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_render.R
\name{normalize_report_format}
\alias{normalize_report_format}
\title{Normalize report output format}
\usage{
normalize_report_format(output_format = "html")
}
\arguments{
\item{output_format}{User-provided output format.}
}
\value{
One of "html" or "docx".
}
\description{
Normalize report output format
}
\keyword{internal}
