% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_core.R
\name{prepare_haiicu_units}
\alias{prepare_haiicu_units}
\title{Prepare combined ICU unit records}
\usage{
prepare_haiicu_units(unit, unit_light)
}
\arguments{
\item{unit}{Standard protocol unit table.}

\item{unit_light}{Light protocol unit table.}
}
\value{
Combined unit table with country/network recoding and global IDs.
}
\description{
Prepare combined ICU unit records
}
