% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_core.R
\name{prepare_patient_units}
\alias{prepare_patient_units}
\title{Prepare patient-unit records}
\usage{
prepare_patient_units(patient, unit)
}
\arguments{
\item{patient}{Patient table.}

\item{unit}{Standard protocol unit table.}
}
\value{
Patient table joined to standard unit records.
}
\description{
Prepare patient-unit records
}
