% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiology.R
\name{recode_resistance_isolate}
\alias{recode_resistance_isolate}
\title{Recode AMR isolate labels}
\usage{
recode_resistance_isolate(result_isolate)
}
\arguments{
\item{result_isolate}{Character vector of TESSy \code{ResultIsolate} codes.}
}
\value{
Character vector of AMR isolate group labels.
}
\description{
Recode AMR isolate labels
}
