% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_data.R
\name{report_output_files}
\alias{report_output_files}
\title{Load report objects from annual outputs}
\usage{
report_output_files()
}
\value{
Named character vector mapping report object names to filenames.
}
\description{
Load report objects from annual outputs
}
